fix: refree comment on pca
Section 3.4: It is not true that PCA requires computing the covariance matrix completely. Recall in lecture that we discussed that one can compute the SVD of the data matrix X instead and derive the principal components from this. This reduces the complexity in this case from $O(n^3)$ to $O(nd^2)$ which is significant. It still may be that one would have to do it in batches in this data because of how large n is, but nonetheless the batches could be larger.